pyproject.toml
[project]
name = "hireme-pyscript"
version = "0.1.0"
description = "Pouya's hireme website"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-mail>=0.10.0",
    "gitpython>=3.1.44",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.5",
]
